Update company logo and alternate text
While investigating an issue with the HTML preview of invoices (which is used by the mobile app), I discovered that the wrong parameter was being used to set the company logo.  The $logo variable contained a the local file location and not a URL-compatible path.

At the same time, all invoice templates were updated to have consistent code formatting and the placeholder alternate text of "Company Logo" was replaced with the company name variable.