Remove helm charts and update the CI accordingly
This commit removes the helm charts as they are now ported
to a new repository https://github.com/Bahmni/crater-extensions
which has the custom setup for Bahmni. The CI is also updated
to create and publish crater-php images with the tag 'product'
that will be used a base image in crater-extensions to create
the crater-php image with the tag 'latest'.

Co-authored-by: Venu Vardhan Reddy Tekula <venuvardhan.t@beehyv.com>